Add demos to docs
