+++ docs/demos/custom_element.rst
@@ -1,0 +1,6 @@
+#########################
+Defining a custom element
+#########################
+
+.. literalinclude:: ../../demo/custom_element.py
+  :language: Python
