+++ docs/demos/index.rst
@@ -1,0 +1,9 @@
+############
+Symfem demos
+############
+
+.. toctree::
+   :titlesonly:
+
+   custom_element
+
